refactor: renomeia FiltersBar para CoeqFiltersBar
