+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqFiltersBar/index.tsx
@@ -1,0 +1,1 @@
+export { CoeqFiltersBar as FiltersBarCoeq } from "./container";
