--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/FiltersBar/index.tsx
@@ -1,1 +1,0 @@
-export { FiltersBarCoeq } from "./container";
